+++ CHANGELOG.md
@@ -9,7 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
--   None
+-   Select Conda environment for SageMath
+-   using venv for lsp
+
+### Fixed
+
+-   change package.json devDependencies to dependencies to package and publish the extension
 
 ## [0.0.2] - 2025-06-21